requirements.txt
Pillow>=9.0.0
rasterio>=1.3.0
numpy>=1.22.0
ttkbootstrap>=1.10.0
tkinterdnd2>=0.3.0 